se pasa int a double

+++ app/src/main/java/com/app/gamercalculator/domain/entities/DollarTaxes.kt
@@ -1,3 +1,3 @@
 package com.app.gamercalculator.domain.entities
 
-data class DollarTaxes(val name: String , val date: String , val taxIva: Int,val taxArca: Int, val mountTotal: Int)
+data class DollarTaxes(val name: String , val date: String , val taxIva: Double,val taxArca: Double, val mountTotal: Double)
